small structure changes to allow different string generators
