+++ src/attributes.js
@@ -7,133 +7,6 @@ const {
   smallComparedTo
 } = require('./lib/utils');
 
-
-const nativeAttrs = (js, ele, idDims, childParent) => {
-  const dims = idDims[js.id] ? idDims[js.id] : {};
-  let componentStyles = {};
-
-  const {
-    top,
-    left,
-    width,
-    // height,
-    position,
-    ...rest
-  } = js.style
-
-  let jsStyles = {...rest}
-
-  if(width < 360) {
-    jsStyles.width = width;
-  }
-
-  // Note: 360 is the max width of iphone
-  if(js.type == 'View' && !jsStyles.width && dims.width < 360) {
-    jsStyles.width = dims.width;
-  }
-
-  if(js.type == 'Text' || js.type == 'Tspan') {
-    const parent = childParent[js.id];
-    const parentDims = parent && idDims[parent];
-    if(parentDims) {
-      const spaceBefore = dims.left - parentDims.left;
-      const spaceAfter = parentDims.right - dims.right;
-
-      if(aboutCentered(spaceBefore, spaceAfter) && !aboutZero(spaceBefore)) {
-        jsStyles['textAlign'] = 'center';
-      }
-
-      if(!aboutCentered(spaceBefore, spaceAfter) && spaceBefore < spaceAfter) {
-        jsStyles['textAlign'] = 'left';
-        jsStyles['marginLeft'] = spaceBefore;
-      }
-
-      if(!aboutCentered(spaceBefore, spaceAfter) && spaceBefore > spaceAfter) {
-        jsStyles['textAlign'] = 'right';
-        jsStyles['marginRight'] = spaceAfter;
-      }
-
-    }
-  }
-
-
-  const parent = childParent[ele.id]
-  if(parent) {
-    const parentDims = parent == 'row' || parent == 'column' ? parent.parentDims : idDims[parent]
-    if(js.type == 'View' && !jsStyles.height && !aboutEqual(dims.height, parentDims.height)) {
-      jsStyles.height = dims.height;
-    }
-  }
-
-  let styles = jsStyles ? Object.keys(jsStyles).reduce((r, key) => {
-    if(key == 'top' || 
-      key == 'left' ||
-      key == 'flex' || 
-      key == 'height' || 
-      key == 'width' ||
-      key == 'fontSize' ||
-      key == 'lineHeight' ||
-      key == 'borderRadius' || 
-      key == 'paddingTop' || 
-      key == 'paddingBottom' ||
-      key == 'paddingRight' ||
-      key == 'paddingLeft' ||
-      key == 'marginTop' ||
-      key == 'marginBottom' ||
-      key == 'marginLeft' ||
-      key == 'marginRight') {
-      if(jsStyles[key] > 0) {
-        r[key] = Math.round(jsStyles[key])        
-      }
-    } else {
-      r[key] = `'${jsStyles[key]}'`      
-    }
-    return r;
-  }, {}) : {};
-
-
-  if(ele.alignJustify) {
-    Object.keys(ele.alignJustify).forEach((key) => {
-      if(key == 'paddingTop' || 
-        key == 'paddingBottom' ||
-        key == 'paddingRight' ||
-        key == 'paddingLeft' ||
-        key == 'marginTop' ||
-        key == 'marginBottom' ||
-        key == 'marginLeft' || 
-        key == 'marginRight') {
-        if(ele.alignJustify[key] > 0) {
-          styles[key] = Math.round(ele.alignJustify[key]);          
-        }
-      } else {
-        styles[key] = `'${ele.alignJustify[key]}'`
-      }
-    })
-  }
-
-
-  let attrObjs = {}
-
-  if(js.directAttrs) {
-    Object.keys(js.directAttrs).forEach((key) => {
-      attrObjs[key] = "'" + js.directAttrs[key] + "'"
-    })
-  }
-
-  const attrString = Object.keys(attrObjs).map((key) => {
-    return `${key}={${attrObjs[key]}}`
-  }).join(" ");
-
-  const attrs = attrString.length > 0 ? " " + attrString : "";
-  let styleId = js.id.replace(/[^\_0-9a-zA-Z]/g, '')
-  if(styleId.match(/^[0-9]/)) {
-    styleId = "_" + styleId
-  }
-
-  componentStyles[styleId] = styles;
-  return({attrs, attrObjs, styleId, componentStyles});
-}
-
 const firstBackgroundColor = (js, base, idDims) => {
   const rDims = idDims[base.id]
   const jsDims = idDims[js.id]
@@ -264,7 +137,5 @@ const determineAlignJustify = (comp, children, idDims, jsObjs) => {
   return alignJustify;
 }
 
-
-module.exports.nativeAttrs = nativeAttrs;
 module.exports.determineAlignJustify = determineAlignJustify;
 module.exports.firstBackgroundColor = firstBackgroundColor;
