+++ src/targets/index.js
@@ -1,0 +1,7 @@
+const reactNative = require('./react-native');
+const react = require('./react');
+
+module.exports = {
+  react,
+  "react-native": reactNative
+};
